reorganization of notebooks, mock validation testing with skylines, generate forecast with contaminants
