+++ cupix/parameter_inference/posterior.py
