+++ notebooks/README.md
@@ -1,0 +1,7 @@
+The notebooks are organized as follows:
+
+Analysis: notebooks used in the DR2 analysis to generate figures, tables, key tests, or key results for the DR2 papers.
+Development: notebooks used to develop cupix, e.g. testing comparing different versions of the code and ensuring good code timing and accurate performance
+Tutorials: example notebooks for new users, walking through different use cases of cupix.
+
+For developers: notebooks in Tutorials should be kept up-to-date for new PRs.
