+++ notebooks/development/average_theory.py
@@ -28,9 +28,9 @@ from lace.cosmo import cosmology
 from cupix.px_data.data_DESI_DR2 import DESI_DR2
 from cupix.likelihood.theory import Theory
 from cupix.likelihood.likelihood import Likelihood
-from cupix.likelihood.posterior import Posterior
+from cupix.parameter_inference.posterior import Posterior
 from cupix.likelihood.free_parameter import FreeParameter
-from cupix.sampling.minimize_posterior import Minimizer
+from cupix.parameter_inference.minimize_posterior import Minimizer
 
 # %% [markdown]
 # ### Read the Px from the stack of 50 mocks
