+++ notebooks/development/model_hcd_lucia.py
