+++ notebooks/development/test_old_minimizer.py
